refactor(extract): the depth guard moves into its own module
html.rs stood at 1390 of its 1500-line cap, and the follow-ups the review
asked for (per-arm tests for closes_implicitly, a noscript fast-path test,
a parser-contract pin) add ~150 lines to it. Carve the pre-parse depth
guard — the cap, the byte scan, the tag-scanning primitives and every one
of their tests — into html/depth_guard.rs, verbatim: no logic moved, no
signature changed, no public API touched.

  html.rs         1390 -> 482 lines
  depth_guard.rs             936 lines (new)

guard_depth is re-exported from html so lib.rs:153 and digest.rs:85 keep
their call sites; the cap is re-exported under #[cfg(test)] because only
the callers' depth-bomb tests read it (a plain re-export is an unused
import in a non-test build).

Evidence at this commit:
  cargo test -p nika-extract --lib   148 passed; 0 failed (148 before)
  cargo clippy -p nika-extract --all-targets -- -D warnings   exit 0
  cargo fmt -p nika-extract -- --check                        exit 0
  scripts/ci/check-fn-length.sh   OK  all fns <= 100 lines
  scripts/ci/check-loc-limits.sh  OK  all .rs files <= 1500 LOC
